fix: users default scope & disallow user_account deletion
the deal with deleting user_accounts is that the default scope for User
is restricted to the existing user_accounts. so, if one is deleted
existing resources associated to it will not be correctly displayed
and select filters won't show those users.

so, user account deletion is disallowed, and the default scope
for account owners is not restricted to active user accounts anymore